0.2.0 release
