+++ app/backend/paprocessing/app.py
@@ -1,5 +1,4 @@
-# Filename: pa_processing_api.py
-
+# FastAPI wrapper for full end-to-end PA processing -> Result Rest API
 import os
 import logging
 import time
