--- app/backend/paprocessing/requirements.txt
@@ -1,26 +1,0 @@
-fastapi==0.104.1
-pydantic==2.5.2
-pydantic_core==2.14.5
-httpx==0.25.2
-python-dotenv==1.0.0
-uvicorn==0.24.0.post1
-requests==2.31.0
-azure-identity
-azure-search-documents==11.4.0
-azure-storage-blob
-bs4
-langchain
-langchain-community
-msal>=0.6.1,<2
-openai
-nest_asyncio
-azure-ai-documentintelligence
-pandas
-pypdf
-azure-cognitiveservices-speech
-PyPDF2
-python-docx
-tiktoken
-spacy
-azure-ai-formrecognizer
-docx2txt
