+++ src/pipeline/autoDetermination/settings.yaml
@@ -1,0 +1,27 @@
+run:
+  logging:
+    level: "INFO"
+    name: "autoDetermination"
+    enable_tracing: true
+
+4o_autoDetermination:
+  max_tokens: 2048
+  top_p: 0.85
+  temperature: 0.7
+  frequency_penalty: 0.0
+  presence_penalty: 0.0
+  system_prompt: "prior_auth_user_prompt.jinja"
+  user_prompt: "prior_auth_user_prompt.jinja"
+  use_o1: False
+
+o1_autoDetermination:
+  max_completion_tokens: 15000
+  user_prompt: "prior_auth_o1_user_prompt.jinja"
+  use_o1: True
+
+
+
+
+
+
+
